test: more 18 test
